Move tree translations to hooks
