+++ packages/@react-aria/tree/intl/ar-AE.json
@@ -1,0 +1,4 @@
+{
+  "expand": "Expand",
+  "collapse": "Collapse"
+}
