+++ packages/@react-aria/tree/intl/en-US.json
@@ -1,0 +1,4 @@
+{
+  "expand": "Expand",
+  "collapse": "Collapse"
+}
