+++ packages/@react-aria/tree/package.json
@@ -23,8 +23,11 @@
   },
   "dependencies": {
     "@react-aria/gridlist": "^3.7.3",
+    "@react-aria/i18n": "^3.10.2",
     "@react-aria/selection": "^3.17.3",
+    "@react-aria/utils": "^3.23.2",
     "@react-stately/tree": "^3.7.5",
+    "@react-types/button": "^3.9.2",
     "@react-types/shared": "^3.22.0",
     "@swc/helpers": "^0.5.0"
   },
