+++ packages/@react-aria/tree/src/useTreeGridListItem.ts
@@ -10,10 +10,15 @@
  * governing permissions and limitations under the License.
  */
 
+import {AriaButtonProps} from '@react-types/button';
 import {AriaGridListItemOptions, GridListItemAria, useGridListItem} from '@react-aria/gridlist';
 import {DOMAttributes, FocusableElement, Node} from '@react-types/shared';
+// @ts-ignore
+import intlMessages from '../intl/*.json';
+import {isAndroid} from '@react-aria/utils';
 import {RefObject} from 'react';
 import {TreeState} from '@react-stately/tree';
+import {useLocalizedStringFormatter} from '@react-aria/i18n';
 
 export interface AriaTreeGridListItemOptions extends Omit<AriaGridListItemOptions, 'isVirtualized'> {
   /** An object representing the treegrid item. Contains all the relevant information that makes up the treegrid row. */
@@ -25,8 +30,10 @@ export interface TreeGridListItemAria extends GridListItemAria {
   rowProps: DOMAttributes,
   /** Props for the tree grid cell element within the tree grid list row. */
   gridCellProps: DOMAttributes,
-  /** Props for the list item description element, if any. */
-  descriptionProps: DOMAttributes
+  /** Props for the tree grid row description element, if any. */
+  descriptionProps: DOMAttributes,
+  /** Props for the tree grid row expand button. */
+    expandButtonProps: AriaButtonProps
 }
 
 /**
@@ -36,6 +43,25 @@ export interface TreeGridListItemAria extends GridListItemAria {
  * @param ref - The ref attached to the row element.
  */
 export function useTreeGridListItem<T>(props: AriaTreeGridListItemOptions, state: TreeState<T>, ref: RefObject<FocusableElement>): TreeGridListItemAria {
-   // TODO: should it return a state specifically for isExpanded? Or is aria attribute sufficient?
-  return useGridListItem(props, state, ref);
+  let gridListAria = useGridListItem(props, state, ref);
+  let isExpanded = gridListAria.rowProps['aria-expanded'] === true;
+  let stringFormatter = useLocalizedStringFormatter(intlMessages, 'react-aria-components');
+
+  let expandButtonProps = {
+    onPress: () => {
+      if (!gridListAria.isDisabled) {
+        state.toggleKey(props.node.key);
+      }
+    },
+    'aria-label': isExpanded ? stringFormatter.format('collapse') : stringFormatter.format('expand'),
+    // TODO: the below actually isn't enough to have keyboard navigation skip over it, we need it to be a span type button but
+    // RAC buttons are always a "button" type element.
+    tabIndex: isAndroid() ? -1 : null
+  };
+
+  // TODO: should it return a state specifically for isExpanded? Or is aria attribute sufficient?
+  return {
+    ...gridListAria,
+    expandButtonProps
+  };
 }
