+++ packages/react-aria-components/intl/ar-AE.json
@@ -1,7 +1,5 @@
 {
   "dropzoneLabel": "DropZone",
   "selectPlaceholder": "حدد عنصرًا",
-  "tableResizer": "أداة تغيير الحجم",
-  "expand": "Expand",
-  "collapse": "Collapse"
+  "tableResizer": "أداة تغيير الحجم"
 }
